% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{crossoverData}
\alias{crossoverData}
\title{Example Crossover Data}
\format{
\link[GenomicRanges:GRanges-class]{GRanges} objects stored as RDS
files with metadata columns:
\describe{
\item{childId}{Child identifier}
\item{familyId}{Family identifier}
\item{startSnp}{SNP name at crossover interval start}
\item{finishSnp}{SNP name at crossover interval end}
\item{startCm}{Genetic position at interval start (cM)}
\item{finishCm}{Genetic position at interval end (cM)}
}
}
\value{
No R object is returned. Access files via
\code{system.file("extdata", package = "inferRecom")}.
}
\description{
Pre-computed maternal and paternal crossover events from the simCEU dataset.
Used for testing and demonstrating haplotype phasing functionality.
}
\details{
Two crossover datasets are provided:
\itemize{
\item \code{xoMat.rds} - Maternal crossover events
\item \code{xoPat.rds} - Paternal crossover events
}

These objects contain detected crossover events from 3-child families
in the simCEU dataset. Crossovers were identified using:

\preformatted{
# Maternal crossovers
xoMat <- xoDetect(
  plinkFile = "simCEU",
  mapFile = "female_chr4.txt",
  familySize = 3,
  parent = "mother",
  snpFilter = 5,
  cmFilter = 1
)

# Paternal crossovers
xoPat <- xoDetect(
  plinkFile = "simCEU",
  mapFile = "male_chr4.txt",
  familySize = 3,
  parent = "father",
  snpFilter = 5,
  cmFilter = 1
)
}

Access the crossover data using:
\preformatted{
dataPath <- system.file("extdata", package = "inferRecom")
xoMat <- readRDS(file.path(dataPath, "xoMat.rds"))
xoPat <- readRDS(file.path(dataPath, "xoPat.rds"))
}
}
