% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.addHomozygousSNPs}
\alias{.addHomozygousSNPs}
\title{Add Homozygous SNPs from Other Parent}
\usage{
.addHomozygousSNPs(actgMatrixHom, actgMatrixHet, genotypeDf, bim, parent)
}
\arguments{
\item{actgMatrixHom}{Matrix for parent being imputed}

\item{actgMatrixHet}{Matrix from other parent (het SNPs)}

\item{genotypeDf}{Original genotype data frame}

\item{bim}{BIM data frame}

\item{parent}{"father" or "mother"}
}
\value{
Combined data frame with het and hom SNPs
}
\description{
Add Homozygous SNPs from Other Parent
}
\keyword{internal}
