% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.addNonInformativeSNPs}
\alias{.addNonInformativeSNPs}
\title{Add Non-Informative SNPs to Phased Output}
\usage{
.addNonInformativeSNPs(
  phasedData,
  genoSubset,
  bim,
  fatherId,
  motherId,
  fatherCol,
  motherCol
)
}
\arguments{
\item{phasedData}{Data frame with phased informative SNPs}

\item{genoSubset}{Full genotype matrix for the family}

\item{bim}{BIM data frame with SNP information}

\item{fatherId}{Father identifier}

\item{motherId}{Mother identifier}

\item{fatherCol}{Column index for father in genoSubset}

\item{motherCol}{Column index for mother in genoSubset}
}
\value{
Data frame with all SNPs (informative and non-informative)
}
\description{
Adds SNPs that were not informative for phasing back into the output with
genotypes inferred from parental homozygous states where possible. For cases
where one parent is homozygous and one is heterozygous, uses haplotype
information from neighboring informative SNPs to infer phase.
}
\keyword{internal}
