% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.alignSnpData}
\alias{.alignSnpData}
\title{Align SNP Data to Common Set}
\usage{
.alignSnpData(snpDf, allSnps)
}
\arguments{
\item{snpDf}{Data frame with SNP information (must have rsID column)}

\item{allSnps}{Character vector of all SNPs to include}
}
\value{
Data frame with all SNPs aligned
}
\description{
Ensures a data frame contains all specified SNPs in order, filling missing
SNPs with NA values
}
\keyword{internal}
