% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.assignAlleleCodes}
\alias{.assignAlleleCodes}
\title{Assign Allele Codes for One Parent}
\usage{
.assignAlleleCodes(genoDf, parent, fatherCol, motherCol)
}
\arguments{
\item{genoDf}{Informative genotype matrix}

\item{parent}{"father" or "mother"}

\item{fatherCol}{Column index for father}

\item{motherCol}{Column index for mother}
}
\value{
Data frame with allele codes
}
\description{
Determines which allele (1 or 2) was inherited from specified parent
}
\keyword{internal}
