% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.checkSameChromosome}
\alias{.checkSameChromosome}
\title{Check if all chromosomes in GRanges are the same}
\usage{
.checkSameChromosome(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
Logical; TRUE if all chromosomes are identical
}
\description{
Check if all chromosomes in GRanges are the same
}
\keyword{internal}
