% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.convertToNucleotides}
\alias{.convertToNucleotides}
\title{Convert Allele Codes to Nucleotides}
\usage{
.convertToNucleotides(alleleCodeDf, bim)
}
\arguments{
\item{alleleCodeDf}{Data frame with allele codes (1 or 2)}

\item{bim}{BIM data frame with allele information}
}
\value{
Data frame with nucleotide alleles
}
\description{
Convert Allele Codes to Nucleotides
}
\keyword{internal}
