% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.convertToSummarizedExperiment}
\alias{.convertToSummarizedExperiment}
\title{Convert Phased List to SummarizedExperiment}
\usage{
.convertToSummarizedExperiment(phaseList, bim)
}
\arguments{
\item{phaseList}{Named list of phased DataFrames}

\item{bim}{BIM data frame}
}
\value{
SummarizedExperiment object
}
\description{
Convert Phased List to SummarizedExperiment
}
\keyword{internal}
