% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.firstNonNA}
\alias{.firstNonNA}
\title{Get First Non-NA Value}
\usage{
.firstNonNA(x)
}
\arguments{
\item{x}{Vector}
}
\value{
First non-NA value
}
\description{
Get First Non-NA Value
}
\keyword{internal}
