% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hzRun.R
\name{.getGeneticPosition}
\alias{.getGeneticPosition}
\title{Get Genetic Position from Physical Position}
\usage{
.getGeneticPosition(posBp, geneticMap)
}
\arguments{
\item{posBp}{Numeric vector of physical positions in base pairs}

\item{geneticMap}{Data frame with chr, pos and cM columns}
}
\value{
Numeric vector of genetic positions in centiMorgans
}
\description{
Maps physical position (bp) to genetic position (cM)
}
\keyword{internal}
