% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.getInformativeGenotypes}
\alias{.getInformativeGenotypes}
\title{Get Informative Genotypes for Phasing}
\usage{
.getInformativeGenotypes(genoSubset)
}
\arguments{
\item{genoSubset}{Genotype matrix for one family}
}
\value{
Matrix of informative genotypes or NULL
}
\description{
Filters for informative SNPs and checks Mendelian consistency
}
\keyword{internal}
