% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.getOtherAllele}
\alias{.getOtherAllele}
\title{Get Complementary Allele}
\usage{
.getOtherAllele(snpId, allele, bim)
}
\arguments{
\item{snpId}{SNP identifier}

\item{allele}{Current allele (A, C, G, or T)}

\item{bim}{BIM data frame}
}
\value{
Complementary allele
}
\description{
Get Complementary Allele
}
\keyword{internal}
