% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hzRun.R
\name{.getSnpLocations}
\alias{.getSnpLocations}
\title{Get Physical Positions for SNPs}
\usage{
.getSnpLocations(snpNames, mapData)
}
\arguments{
\item{snpNames}{Character vector of SNP names}

\item{mapData}{Data frame with SNP map information}
}
\value{
Numeric vector of positions
}
\description{
Map SNP names to physical positions
}
\keyword{internal}
