% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.granges2DataFrame}
\alias{.granges2DataFrame}
\title{Convert GRanges to DataFrame}
\usage{
.granges2DataFrame(granges)
}
\arguments{
\item{granges}{GRanges object}
}
\value{
data.frame with genomic ranges information and metadata columns
}
\description{
Converts a GRanges object to a data.frame/DataFrame for internal processing
}
\keyword{internal}
