% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.inferHetParentPhase}
\alias{.inferHetParentPhase}
\title{Infer Heterozygous Parent Phase}
\usage{
.inferHetParentPhase(
  neighbors,
  phasedData,
  hetParent,
  fatherId,
  motherId,
  allele1,
  allele2
)
}
\arguments{
\item{neighbors}{List with left and right neighbor indices}

\item{phasedData}{Data frame with phased informative SNPs}

\item{hetParent}{"father" or "mother" - which parent is heterozygous}

\item{fatherId}{Father identifier}

\item{motherId}{Mother identifier}

\item{allele1}{Reference allele}

\item{allele2}{Alternate allele}
}
\value{
List with hap1 and hap2, or NULL if cannot infer
}
\description{
Infers the phase of the heterozygous parent using neighboring informative SNPs
}
\keyword{internal}
