% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.matchAlleleCode}
\alias{.matchAlleleCode}
\title{Match Allele Code for Single Genotype}
\usage{
.matchAlleleCode(parent, fatherGt, motherGt, childGt)
}
\arguments{
\item{parent}{"father" or "mother"}

\item{fatherGt}{Father genotype (0, 1, or 2)}

\item{motherGt}{Mother genotype (0, 1, or 2)}

\item{childGt}{Child genotype (0, 1, or 2)}
}
\value{
Allele code (1 or 2)
}
\description{
Match Allele Code for Single Genotype
}
\keyword{internal}
