% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.phaseChildren}
\alias{.phaseChildren}
\title{Phase Child Haplotypes}
\usage{
.phaseChildren(actgCombPat, actgCombMat)
}
\arguments{
\item{actgCombPat}{Paternal allele data frame}

\item{actgCombMat}{Maternal allele data frame}
}
\value{
Data frame with phased children
}
\description{
Phase Child Haplotypes
}
\keyword{internal}
