% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.phaseParent}
\alias{.phaseParent}
\title{Phase Parent Haplotypes Using Crossover Information}
\usage{
.phaseParent(actgMatrix, xoDetectMatrix, bim, parentId)
}
\arguments{
\item{actgMatrix}{Allele matrix for parent}

\item{xoDetectMatrix}{Crossover detection results}

\item{bim}{BIM data frame}

\item{parentId}{Parent identifier}
}
\value{
Data frame with two phased haplotypes
}
\description{
Phase Parent Haplotypes Using Crossover Information
}
\keyword{internal}
