% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.phaseSingleFamily}
\alias{.phaseSingleFamily}
\title{Phase a Single Family}
\usage{
.phaseSingleFamily(familyId, geno, fam, bim, xoPat, xoMat)
}
\arguments{
\item{familyId}{Family identifier}

\item{geno}{Numeric genotype matrix}

\item{fam}{FAM data frame}

\item{bim}{BIM data frame}

\item{xoPat}{Paternal crossover data frame}

\item{xoMat}{Maternal crossover data frame}
}
\value{
DataFrame with phased haplotypes or character message
}
\description{
Phase a Single Family
}
\keyword{internal}
