% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hzRun.R
\name{.processSingleIndividual}
\alias{.processSingleIndividual}
\title{Process Single Individual for ROH Detection}
\usage{
.processSingleIndividual(
  sampleIdx,
  genoMatrix,
  locationVec,
  snpNames,
  sampleIds,
  minSnps,
  minBp,
  geneticMap,
  minCm
)
}
\arguments{
\item{sampleIdx}{Integer index of sample in genotype matrix}

\item{genoMatrix}{Matrix of genotypes (SNPs x samples)}

\item{locationVec}{Physical positions of SNPs}

\item{snpNames}{SNP identifiers}

\item{sampleIds}{Sample identifiers}

\item{minSnps}{Minimum number of SNPs}

\item{minBp}{Minimum physical length}

\item{geneticMap}{Genetic map data frame or NULL}

\item{minCm}{Minimum genetic length or NULL}
}
\value{
Matrix with ROH information, or NULL
}
\description{
Wrapper function for parallel processing of a single individual
}
\keyword{internal}
