% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoPhase.R
\name{.writeVcfFiles}
\alias{.writeVcfFiles}
\title{Write Phased Genotypes to VCF Files}
\usage{
.writeVcfFiles(phaseList, bim, vcfOutput)
}
\arguments{
\item{phaseList}{Named list of phased DataFrames}

\item{bim}{BIM data frame}

\item{vcfOutput}{Path prefix for VCF output}
}
\value{
Character vector of created VCF file paths
}
\description{
Write Phased Genotypes to VCF Files
}
\keyword{internal}
