% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{simCEU-data}
\alias{simCEU-data}
\title{Example PLINK Dataset - simCEU}
\format{
PLINK binary format files (.bed, .bim, .fam) containing:
\describe{
\item{Samples}{Simulated family trios and larger pedigrees with 2-3 children}
\item{SNPs}{Chromosome 4 markers}
\item{Genotypes}{Simulated genotype data with realistic LD structure}
\item{Families}{Mix of 2-child and 3-child families for testing}
}
}
\source{
Simulated data based on CEU population structure from 1000 Genomes
Project and R package \code{sim1000G}
}
\description{
A simulated CEU (Utah residents with Northern and Western European ancestry)
dataset for demonstrating crossover detection and ROH analysis.
}
\details{
The dataset includes three PLINK binary files:
\itemize{
\item \code{simCEU.bed} - Binary genotype data
\item \code{simCEU.bim} - Variant information (SNP IDs, positions, alleles)
\item \code{simCEU.fam} - Sample information (family IDs, relationships)
}

Access the files using:
\preformatted{
dataPath <- system.file("extdata", package = "inferRecom")
plinkFile <- file.path(dataPath, "simCEU")
}
}
\references{
Siva, N. (2008). 1000 Genomes project.

Dimitromanolakis, A., Xu, J., Krol, A., & Briollais, L. (2019). sim1000G:
a user-friendly genetic variant simulator in R for unrelated individuals and
family-based designs. BMC bioinformatics, 20(1), 26.
}
